Restructured and Optimized the code.
