+++ lib/focus_point.dart
@@ -1,20 +1,16 @@
-import 'package:advanced_graphview/node_module.dart';
+import 'package:advanced_graphview/advanced_graphview_flame.dart';
 import 'package:advanced_graphview/world_map.dart';
 import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
-import 'package:flame/experimental.dart';
 import 'package:flame/game.dart';
-import 'package:flame/input.dart';
 import 'package:flame/palette.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart' as mat;
 import 'package:flutter/services.dart';
 
-import 'advanced_graphview.dart';
-
-class Ember<T extends FlameGame> extends PositionComponent
+class FocusPoint<T extends FlameGame> extends PositionComponent
     with HasGameReference<T> {
-  Ember({super.position, Vector2? size, super.priority, super.key})
+  FocusPoint({super.position, Vector2? size, super.priority, super.key})
       : super(
           size: size ?? Vector2.all(50),
           anchor: Anchor.center,
@@ -25,99 +21,7 @@ class Ember<T extends FlameGame> extends PositionComponent
   Future<void> onLoad() async {}
 }
 
-class FollowComponentExample extends FlameGame
-    with HasCollisionDetection, HasKeyboardHandlerComponents, HasGameRef {
-  FollowComponentExample({
-    required this.nodePadding,
-    required this.nodeSize,
-    required this.graphDataStructure,
-    required this.context,
-    required this.isDebug,
-    required this.onDrawLine,
-    required this.advancedGraphviewController,
-    required this.flameBackgroundColor,
-    required this.pixelRatio,
-    required this.onNodeTap,
-  }) : super(
-          camera: CameraComponent(),
-        );
-
-  @override
-  bool get debugMode => isDebug;
-  final double? pixelRatio;
-  final BuildContext context;
-  late MovableEmber player;
-  final double nodeSize;
-  final double nodePadding;
-  final GraphDataStructure graphDataStructure;
-  final bool isDebug;
-  final Paint? Function(GraphNode lineFrom, GraphNode lineTwo)? onDrawLine;
-  final AdvancedGraphviewController? advancedGraphviewController;
-  final Color? flameBackgroundColor;
-  final Function(GraphNode)? onNodeTap;
-  @override
-  Color backgroundColor() {
-    return flameBackgroundColor ?? mat.Colors.white;
-  }
-
-  @override
-  Future<void> onLoad() async {
-    world.add(WorldMap());
-    world.add(player = MovableEmber());
-
-    //world.add(CirclingEnemy(player: ember));
-
-    // world.add(SimpleBullet(player: ember));
-    camera.setBounds(Rectangle.fromPoints(
-        Vector2(0, 0), Vector2(WorldMap.size, WorldMap.size)));
-    camera.follow(
-      player,
-    );
-    void renderData() {
-      double i = 0;
-      double j = 0;
-      graphDataStructure.generateItem((node) {
-        double nodei = i;
-        double nodej = j;
-        world.add(
-          NodeModule(
-              nodeSize: Vector2((nodeSize + (nodePadding * 2)),
-                  (nodeSize + (nodePadding * 2))),
-              nodePosition: Vector2(nodej, nodei),
-              nodePadding: nodePadding,
-              graphNode: node,
-              nodeImageSize: nodeSize,
-              graphDataStructure: graphDataStructure),
-        );
-
-        if (j < WorldMap.size - (nodeSize + (nodePadding * 2))) {
-          j = j + (nodeSize + (nodePadding * 2));
-        } else {
-          i = i + (nodeSize + (nodePadding * 2));
-          j = 0;
-        }
-      });
-    }
-
-    void addLines() {
-      graphDataStructure.generateItem((node) {
-        world.add(LineDrawer(graphNode: node));
-      });
-    }
-
-    renderData();
-    addLines();
-  }
-
-  static const speed = 2000.0;
-  @override
-  void update(double dt) {
-    super.update(dt);
-    camera.viewfinder.zoom = advancedGraphviewController?.zoom ?? 1;
-  }
-}
-
-class MovableEmber extends Ember<FollowComponentExample>
+class FocusPointImpl extends FocusPoint<AdvancedGraphviewFlame>
     with CollisionCallbacks, KeyboardHandler {
   static const double speed = 300;
   static final TextPaint textRenderer = TextPaint(
@@ -131,7 +35,7 @@ class MovableEmber extends Ember<FollowComponentExample>
   late final minPosition = Vector2.zero() + Vector2.all(25);
   // late BallBorder ballBorder;
   int consumed = 0;
-  MovableEmber() : super(priority: 2);
+  FocusPointImpl() : super(priority: 2);
 
   @override
   Future<void> onLoad() async {
