+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: advanced_graphview
-description: "Advanced Graphview creates best graph in flutter."
-version: 0.0.5
+description: "Advanced Graphview enables users to create Tree, Graph, Topology UI with simple setup. This package can be used to display different flows."
+version: 0.0.6
 homepage: https://github.com/nikhilishwarkulal/advanced_graphview
 
 environment:
